+++ components/Community/Community.js
@@ -28,20 +28,25 @@ const Community = props => {
     getAccount(name).then(profile => {
       if (profile.cover_image && !coverImage)
         setCoverImage(profile.cover_image);
+      else
+        setCoverImage(
+          'https://steemitimages.com/p/vM1pGHgNcyCbeeHc87msUy15a6gssa8TxcCWDULWJwXbLzzTNUCuGN3v6fTG1Cxa3iA7dj2rixvqQw5FRuewtFGmCt4r68CcjssuoSnko5jXFk1x2hcHhcFRVCJFWE3Jf3rfCez',
+        );
     });
   }, []);
 
   return (
     <>
-      <Head shorttitle={title} image={coverImage} description={about} />
+      <Head title={title} image={coverImage} description={about} />
       <Header active="community" subheader={title} />
       <PhotoHeader
-        noEdit
         hideButtons
+        communityTag
         title={title}
         description={about}
         image={coverImage}
         query={{}}
+        tag={name}
       />
       <div className="pb-3 container-fluid" id="containerInvisibleOnMobile">
         <div className="row">
@@ -70,10 +75,10 @@ const Community = props => {
                       {description.replace(
                         'This community is part of TravelFeed. TravelFeed, the number one travel community on Steem, has stepped up the game with the launch of our own platform TravelFeed.io back in June, 2019. We curate all posts from this community, however, you can double your upvotes by posting through TravelFeed.io.',
                         '',
-                      )}
+                      ) || about}
                     </Typography>
                     <Divider className="mt-3 mb-3" />
-                    <div className="container">
+                    <div className="container pb-2">
                       <div className="row">
                         <div className="col-6 text-center">
                           <Typography variant="h4">{subscribers}</Typography>{' '}
